old/back/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
google-cloud-storage==3.4.0
google-cloud-logging==3.11.3
gunicorn==23.0.0
python-dotenv==1.1.1
pandas==2.3.2
pyarrow==21.0.0
openpyxl==3.1.2